+++ Source/Components/List/ListComponentCell.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+class ListComponentCell: UITableViewCell {
+
+  override init(style: UITableViewCellStyle, reuseIdentifier: String!) {
+    super.init(style: .Subtitle, reuseIdentifier: reuseIdentifier)
+  }
+
+  required init?(coder aDecoder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }  
+}
