+++ Source/Library/ComponentSizeDelegate.swift
@@ -1,0 +1,3 @@
+protocol ComponentSizeDelegate: class {
+  func sizeDidUpdate()
+}
